requirements.txt
django
djangorestframework==3.8.2
psycopg2
pyYAML
django-tastypie == 0.14.0

django-extensions
